# access_control_service/access-control-service-openapi.yaml
openapi: 3.0.3
info:
  title: Compliance & Access Control Service API
  description: |
    Микросервис для управления геозонами, правилами доступа и контроля соблюдения политик безопасности.
    Обеспечивает привязку меток к сущностям (сотрудники/оборудование), проверку разрешенных зон и генерацию оповещений о нарушениях.
  version: 1.0.0
  contact:
    name: Команда разработки
    email: security@rtls.example.com
servers:
  - url: http://localhost:8084/api/v1
    description: Локальный сервер разработки
  - url: https://compliance.rtls.example.com/api/v1
    description: Продуктивный сервер
...
paths:
...
  /geofences:
    get:
      tags:
        - Geofences
      summary: Получение списка всех геозон
      description: Возвращает список всех зарегистрированных геозон.
      operationId: getAllGeofences
      responses:
        '200':
          description: Успешный запрос.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Geofence'
        '500':
          description: Ошибка сервера.

    post:
      tags:
        - Geofences
      summary: Создание новой геозоны
      description: Создает новую геозону (зону безопасности, доступа и т.д.).
      operationId: createGeofence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeofenceCreate'
            examples:
              rectangularZone:
                summary: Прямоугольная зона
                value:
                  name: "Серверная комната"
                  zone_type: "restricted"
                  description: "Запрещенная зона для посторонних"
                  shape: "rectangle"
                  coordinates:
                    min_x: 10.0
                    max_x: 15.0
                    min_y: 5.0
                    max_y: 10.0
                    min_z: 0.0
                    max_z: 3.0
              circularZone:
                summary: Круговая зона
                value:
                  name: "Зона безопасности станка"
                  zone_type: "danger"
                  description: "Опасная зона вокруг станка"
                  shape: "circle"
                  coordinates:
                    center_x: 25.0
                    center_y: 30.0
                    radius: 3.0
      responses:
        '201':
          description: Геозона успешно создана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        '400':
          description: Некорректные данные геозоны.
  /geofences/{geofence_id}:
    get:
      tags:
        - Geofences
      summary: Получение информации о конкретной геозоне
      description: Возвращает детальную информацию о геозоне.
      operationId: getGeofenceById
      parameters:
        - name: geofence_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешный запрос.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        '404':
          description: Геозона с указанным ID не найдена.

    put:
      tags:
        - Geofences
      summary: Обновление геозоны
      description: Полностью обновляет информацию о геозоне.
      operationId: updateGeofence
      parameters:
        - name: geofence_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeofenceCreate'
      responses:
        '200':
          description: Геозона успешно обновлена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        '404':
          description: Геозона с указанным ID не найдена.

    delete:
      tags:
        - Geofences
      summary: Удаление геозоны
      description: Удаляет геозону из системы.
      operationId: deleteGeofence
      parameters:
        - name: geofence_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Геозона успешно удалена.
        '404':
          description: Геозона с указанным ID не найдена.
  /geofences/check:
    post:
      tags:
        - Geofences
      summary: Проверка нахождения точки в геозонах
      description: Проверяет, находится ли указанная точка внутри какой-либо геозоны.
      operationId: checkPointInGeofences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointCheckRequest'
      responses:
        '200':
          description: Успешная проверка.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeofenceCheckResult'

  # ==================== Rules: Управление правилами ====================
...
components:
  schemas:
...
    GeofenceCreate:
      type: object
      required:
        - name
        - zone_type
        - shape
        - coordinates
      properties:
        name:
          type: string
          description: Название геозоны
        zone_type:
          type: string
          enum: [restricted, danger, safe, work_area, parking, other]
          description: Тип зоны
        description:
          type: string
          description: Описание зоны
        shape:
          type: string
          enum: [rectangle, circle, polygon]
          description: Форма зоны
        coordinates:
          type: object
          description: Координаты зоны (формат зависит от shape)
          properties:
            # Для rectangle
            min_x:
              type: number
            max_x:
              type: number
            min_y:
              type: number
            max_y:
              type: number
            min_z:
              type: number
            max_z:
              type: number
            # Для circle
            center_x:
              type: number
            center_y:
              type: number
            radius:
              type: number
            # Для polygon
            vertices:
              type: array
              items:
                type: object
                properties:
                  x:
                    type: number
                  y:
                    type: number
        buffer_meters:
          type: number
          default: 0
          description: Буферная зона вокруг геозоны (в метрах)
        is_active:
          type: boolean
          default: true
    Geofence:
      allOf:
        - $ref: '#/components/schemas/GeofenceCreate'
        - type: object
          properties:
            geofence_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    PointCheckRequest:
      type: object
      required:
        - x
        - y
      properties:
        x:
          type: number
          format: float
        y:
          type: number
          format: float
        z:
          type: number
          format: float
          default: 0.0
        geofence_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Список ID геозон для проверки (если не указано - проверяются все)
    GeofenceCheckResult:
      type: object
      properties:
        point:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            z:
              type: number
        intersections:
          type: array
          items:
            type: object
            properties:
              geofence_id:
                type: string
                format: uuid
              geofence_name:
                type: string
              zone_type:
                type: string
              is_inside:
                type: boolean

    # --- Правила ---
...